service增加了注解

+++ privilegeservice/src/main/java/cn/edu/xmu/privilegegateway/privilegeservice/model/vo/PrivilegeRetVo.java
@@ -19,7 +19,7 @@ public class PrivilegeRetVo {
 
     private String url;
 
-    private Integer requestType;
+    private Byte requestType;
 
     private LocalDateTime gmtCreate;
